EcoMarket-Compartir1 - Semana 9/requirements.txt
fastapi
uvicorn[standard]
pika
pydantic
pydantic-settings
email-validator
psycopg2-binary
asyncpg
python-dotenv
pyjwt
pytest
pytest-asyncio
httpx
cryptography
